+++ coverage.txt
@@ -21,7 +21,7 @@
           √ debugMessage logt nicht bei deaktivierten Logging
           √ infoMessage logt nicht bei deaktivierten Logging
           √ logMessage logt nicht bei deaktivierten Logging
-      13 passing (10ms)
+      13 passing (9ms)
     ----------------------------------|---------|----------|---------|---------|-------------------
     File                              | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s 
     ----------------------------------|---------|----------|---------|---------|-------------------
@@ -43,10 +43,11 @@
      liona-feeds/src                  |       0 |        0 |       0 |       0 |                   
       index.ts                        |       0 |        0 |       0 |       0 |                   
      liona-feeds/src/feeds            |    4.25 |    10.52 |    2.85 |    4.31 |                   
-      cleaner.ts                      |       0 |        0 |       0 |       0 | 17-51             
+      cleaner.ts                      |       0 |        0 |       0 |       0 | 7-41              
       feeder.ts                       |       0 |        0 |       0 |       0 | 13-177            
+      metadata.ts                     |       0 |        0 |       0 |       0 |                   
       ranking.ts                      |     100 |      100 |     100 |     100 |                   
-      statistic.ts                    |       0 |        0 |       0 |       0 | 20-83             
+      statistic.ts                    |       0 |        0 |       0 |       0 | 14-77             
      liona-feeds/src/network          |       0 |        0 |       0 |       0 |                   
       cors.ts                         |       0 |        0 |       0 |       0 | 2-8               
      liona-feeds/src/shared           |     100 |      100 |     100 |     100 |
